fix: write tests for adapter-ndk * with updating minimum version of NDK
