+++ packages/adapter-ndk/package.json
@@ -28,8 +28,10 @@
   },
   "scripts": {
     "prepack": "yarn build",
-    "check-type": "tsc -p tsconfig.lint.json || exit 0",
-    "lint": "run-p check-type lint:*",
+    "tsc": "tsc",
+    "test": "vitest run",
+    "testw": "vitest",
+    "lint": "run-p tsc lint:*",
     "lint:format": "prettier --check --loglevel warn src/**/*.ts",
     "lint:js": "eslint --max-warnings 0 src/",
     "fix": "run-s fix:*",
@@ -43,10 +45,10 @@
     "@nostr-fetch/kernel": "^0.9.0"
   },
   "devDependencies": {
-    "@nostr-dev-kit/ndk": "^0.4.6"
+    "@nostr-dev-kit/ndk": "^0.5.0"
   },
   "peerDependencies": {
-    "@nostr-dev-kit/ndk": "^0.4.6",
+    "@nostr-dev-kit/ndk": "^0.5.0",
     "nostr-fetch": "^0.9.0"
   }
 }
